+++ tests/testthat/test-provider-claude.R
@@ -62,3 +62,9 @@ test_that("can set beta headers", {
   req <- chat_request(provider)
   expect_equal(req$headers$`anthropic-beta`, c("a", "b"))
 })
+
+test_that("continues to work after whitespace only outputs (#376)", {
+  chat <- chat_claude()
+  chat$chat("Respond with only two blank lines")
+  expect_equal(chat$chat("What's 1+1? Just give me the number"), "2")
+})